added permission check to api routes, rebuilt alembic

+++ app/api/v1/endpoints/clients.py
@@ -1,11 +1,12 @@
-from typing import Dict
+from typing import Any, Dict
 
 from fastapi import APIRouter, Depends
+from sqlalchemy import Select
 
 from app.api.deps import (
-    AsyncDatabaseSession,
-    CurrentUser,
-    FetchClientOr404,
+    CommonUserQueryParams,
+    GetUserQueryParams,
+    Permission,
     PermissionController,
     get_async_db,
     get_client_or_404,
@@ -14,22 +15,27 @@ from app.api.deps import (
 )
 from app.api.exceptions import ClientAlreadyExists
 from app.api.openapi import clients_read_responses
-from app.core.pagination import (
-    GetPaginatedQueryParams,
-    PageParams,
-    PageParamsFromQuery,
-    Paginated,
-)
+from app.core.pagination import PageParams, Paginated
 from app.core.security import auth
 from app.core.security.permissions import (
+    AccessDelete,
+    AccessDeleteSelf,
+    AccessRead,
+    AccessReadRelated,
+    AccessReadSelf,
+    AccessUpdate,
+    AccessUpdateRelated,
+    AccessUpdateSelf,
     RoleAdmin,
     RoleClient,
     RoleEmployee,
     RoleManager,
+    RoleUser,
 )
 from app.crud import ClientRepository
 from app.models import Client
-from app.schemas import ClientCreate, ClientRead, ClientUpdate
+from app.schemas import ClientCreate, ClientDelete, ClientRead, ClientUpdate
+from app.worker import task_request_to_delete_client
 
 router: APIRouter = APIRouter()
 
@@ -38,7 +44,7 @@ router: APIRouter = APIRouter()
     "/",
     name="clients:list",
     dependencies=[
-        Depends(PageParamsFromQuery),
+        Depends(CommonUserQueryParams),
         Depends(auth.implicit_scheme),
         Depends(get_async_db),
         Depends(get_current_user),
@@ -47,8 +53,7 @@ router: APIRouter = APIRouter()
     response_model=Paginated[ClientRead],
 )
 async def clients_list(
-    query: GetPaginatedQueryParams,
-    db: AsyncDatabaseSession,
+    query: GetUserQueryParams,
     permissions: PermissionController = Depends(get_permission_controller),
 ) -> Paginated[ClientRead]:
     """Retrieve a paginated list of clients.
@@ -57,7 +62,7 @@ async def clients_list(
     ------------
     `role=admin|manager` : all clients
 
-    `role=client|employee` : only clients associated with the user via `user_client`
+    `role=user` : only clients associated with the user via `user_client`
         table
 
     Returns:
@@ -65,12 +70,20 @@ async def clients_list(
     `Paginated[ClientRead]` : a paginated list of clients, optionally filtered
 
     """
-    clients_repo: ClientRepository = ClientRepository(session=db)
+    # formulate the select statement based on the current user's role
+    select_stmt: Select
+    if RoleAdmin in permissions.privileges or RoleManager in permissions.privileges:
+        select_stmt = permissions.client_repo.query_list(user_id=query.user_id)
+    else:  # TODO: test
+        select_stmt = permissions.client_repo.query_list(
+            user_id=permissions.current_user.id
+        )
+    # return role based response
     response_out: Paginated[
         ClientRead
     ] = await permissions.get_paginated_resource_response(
         table_name=Client.__tablename__,
-        stmt=clients_repo.query_list(),
+        stmt=select_stmt,
         page_params=PageParams(page=query.page, size=query.size),
         responses={
             RoleAdmin: ClientRead,
@@ -88,13 +101,14 @@ async def clients_list(
     dependencies=[
         Depends(auth.implicit_scheme),
         Depends(get_async_db),
+        Depends(get_current_user),
+        Depends(get_permission_controller),
     ],
     response_model=ClientRead,
 )
 async def clients_create(
-    current_user: CurrentUser,
-    db: AsyncDatabaseSession,
     client_in: ClientCreate,
+    permissions: PermissionController = Depends(get_permission_controller),
 ) -> ClientRead:
     """Create a new client.
 
@@ -107,7 +121,9 @@ async def clients_create(
     `ClientRead` : the newly created client
 
     """
-    clients_repo: ClientRepository = ClientRepository(session=db)
+    # verify current user has permission to take this action
+    await permissions.verify_user_can_access(privileges=[RoleAdmin, RoleManager])
+    clients_repo: ClientRepository = ClientRepository(session=permissions.db)
     data: Dict = client_in.model_dump()
     check_title: str | None = data.get("title")
     if check_title:
@@ -118,7 +134,14 @@ async def clients_create(
         if a_client:
             raise ClientAlreadyExists()
     new_client: Client = await clients_repo.create(client_in)
-    return ClientRead.model_validate(new_client)
+    # return role based response
+    response_out: ClientRead = permissions.get_resource_response(
+        resource=new_client,
+        responses={
+            RoleUser: ClientRead,
+        },
+    )
+    return response_out
 
 
 @router.get(
@@ -128,14 +151,17 @@ async def clients_create(
         Depends(auth.implicit_scheme),
         Depends(get_async_db),
         Depends(get_client_or_404),
+        Depends(get_current_user),
+        Depends(get_permission_controller),
     ],
     responses=clients_read_responses,
     response_model=ClientRead,
 )
 async def clients_read(
-    current_user: CurrentUser,
-    db: AsyncDatabaseSession,
-    client: FetchClientOr404,
+    client: Client = Permission(
+        [AccessRead, AccessReadSelf, AccessReadRelated], get_client_or_404
+    ),
+    permissions: PermissionController = Depends(get_permission_controller),
 ) -> ClientRead:
     """Retrieve a single client by id.
 
@@ -150,7 +176,19 @@ async def clients_read(
     `ClientRead` : a client matching the client_id
 
     """
-    return ClientRead.model_validate(client)
+    # verify current user has permission to take this action
+    await permissions.verify_user_can_access(
+        privileges=[RoleAdmin, RoleManager],
+        client_id=client.id,
+    )
+    # return role based response
+    response_out: ClientRead = permissions.get_resource_response(
+        resource=client,
+        responses={
+            RoleUser: ClientRead,
+        },
+    )
+    return response_out
 
 
 @router.patch(
@@ -160,14 +198,17 @@ async def clients_read(
         Depends(auth.implicit_scheme),
         Depends(get_async_db),
         Depends(get_client_or_404),
+        Depends(get_current_user),
+        Depends(get_permission_controller),
     ],
     response_model=ClientRead,
 )
 async def clients_update(
-    current_user: CurrentUser,
-    db: AsyncDatabaseSession,
-    client: FetchClientOr404,
     client_in: ClientUpdate,
+    client: Client = Permission(
+        [AccessUpdate, AccessUpdateSelf, AccessUpdateRelated], get_client_or_404
+    ),
+    permissions: PermissionController = Depends(get_permission_controller),
 ) -> ClientRead:
     """Update a client by id.
 
@@ -175,12 +216,26 @@ async def clients_update(
     ------------
     `role=admin|manager` : all clients
 
+    `role=user` : only clients associated with the user via `user_client`
+
     Returns:
     --------
     `ClientRead` : the updated client
 
     """
-    clients_repo: ClientRepository = ClientRepository(session=db)
+    # verify the input schema is valid for the current user's role
+    permissions.verify_input_schema_by_role(
+        input_object=client_in,
+        schema_privileges={
+            RoleUser: ClientUpdate,
+        },
+    )
+    # verify current user has permission to take this action
+    await permissions.verify_user_can_access(
+        privileges=[RoleAdmin, RoleManager],
+        client_id=client.id,
+    )
+    clients_repo: ClientRepository = ClientRepository(session=permissions.db)
     if client_in.title is not None:
         a_client: Client | None = await clients_repo.read_by(
             field_name="title", field_value=client_in.title
@@ -190,11 +245,14 @@ async def clients_update(
     updated_client: Client | None = await clients_repo.update(
         entry=client, schema=client_in
     )
-    return (
-        ClientRead.model_validate(updated_client)
-        if updated_client
-        else ClientRead.model_validate(client)
+    # return role based response
+    response_out: ClientRead = permissions.get_resource_response(
+        resource=updated_client if updated_client else client,
+        responses={
+            RoleUser: ClientRead,
+        },
     )
+    return response_out
 
 
 @router.delete(
@@ -204,25 +262,49 @@ async def clients_update(
         Depends(auth.implicit_scheme),
         Depends(get_async_db),
         Depends(get_client_or_404),
+        Depends(get_current_user),
+        Depends(get_permission_controller),
     ],
-    response_model=None,
+    response_model=ClientDelete,
 )
 async def clients_delete(
-    current_user: CurrentUser,
-    db: AsyncDatabaseSession,
-    client: FetchClientOr404,
-) -> None:
+    client: Client = Permission([AccessDelete, AccessDeleteSelf], get_client_or_404),
+    permissions: PermissionController = Depends(get_permission_controller),
+) -> ClientDelete:
     """Delete a client by id.
 
     Permissions:
     ------------
     `role=admin` : all clients
 
+    `role=client` : may request to have their client data deleted
+
     Returns:
     --------
-    `None`
+    `ClientDelete` : a message indicating the user deleted a client or if a user
+        requested to delete a client they are associated with
 
     """
-    clients_repo: ClientRepository = ClientRepository(session=db)
-    await clients_repo.delete(entry=client)
-    return None
+    await permissions.verify_user_can_access(
+        privileges=[RoleAdmin], client_id=client.id
+    )
+    client_delete: ClientDelete
+    if RoleAdmin in permissions.privileges:
+        clients_repo: ClientRepository = ClientRepository(session=permissions.db)
+        await clients_repo.delete(entry=client)
+        client_delete = ClientDelete(
+            message="Client deleted",
+            user_id=permissions.current_user.id,
+            client_id=client.id,
+        )
+    else:  # TODO: test
+        delete_client_task: Any = task_request_to_delete_client.delay(
+            user_id=permissions.current_user.id, client_id=client.id
+        )
+        client_delete = ClientDelete(
+            message="Client requested to be deleted",
+            user_id=permissions.current_user.id,
+            client_id=client.id,
+            task_id=delete_client_task.task_id,
+        )
+    return client_delete
